moving quick open files to search folder
