+++ src/search/QuickOpenCSS.js
