+++ src/search/QuickOpenHTML.js
