+++ src/search/QuickOpenJavaScript.js
